Improve pattern matching MIR lowering
